경비 등록

+++ src/main/java/com/ezkorea/hybrid_app/service/aws/AWSService.java
@@ -7,6 +7,8 @@ import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.ezkorea.hybrid_app.domain.aws.S3Image;
 import com.ezkorea.hybrid_app.domain.aws.S3ImageRepository;
+import com.ezkorea.hybrid_app.domain.expenses.Expenses;
+import com.ezkorea.hybrid_app.domain.expenses.ExpensesRepository;
 import com.ezkorea.hybrid_app.domain.notice.Notice;
 import com.ezkorea.hybrid_app.domain.notice.NoticeRepository;
 import com.ezkorea.hybrid_app.domain.user.member.Member;
@@ -39,6 +41,7 @@ public class AWSService {
     private final S3ImageRepository s3ImageRepository;
     private final NoticeRepository noticeRepository;
     private final MemberRepository memberRepository;
+    private final ExpensesRepository expensesRepository;
     private final MemberService memberService;
 
     /**
@@ -50,41 +53,58 @@ public class AWSService {
     public void saveImageInCurrentEntity(S3ImageDto dto, List<MultipartFile> multipartFileList) {
         StringBuilder sb;
         switch (dto.getEntity()) {
+            // notice : 다중 파일 업로드
             case "notice" -> {
                 Notice currentNotice = noticeRepository.findById(dto.getId())
                         .orElseThrow( () -> new IdNotFoundException("해당하는 id가 존재하지 않습니다."));
-                for (String url : saveImage(dto, multipartFileList)) {
-                    sb = new StringBuilder();
-                    sb.append("images/").append(dto.getEntity()).append("/").append(dto.getId()).append("/");
-                    S3Image savedImage = s3ImageRepository.save(S3Image.builder()
-                            .notice(currentNotice)
-                            .filePath(url)
-                            .fileName(url.split(sb.toString())[1])
-                            .fileRepo(sb.toString())
-                            .build());
-                    currentNotice.getImageList().add(savedImage);
+                for (S3Image s3Image : makeNewS3ImageObject(dto, multipartFileList)) {
+                    s3Image.setNotice(currentNotice);
+                    currentNotice.getImageList().add(s3Image);
                 }
             }
+            // member : 단일 파일 업로드 + 기존 S3Image 객체 수정 방식
             case "member" -> {
                 Member currentMember = memberRepository.findById(dto.getId())
                         .orElseThrow(() -> new IdNotFoundException("해당하는 id가 존재하지 않습니다."));
 
+                // Member는 무조건 프로필 사진이 존재해야하기 때문에 false
                 deleteS3Image(currentMember.getS3Image(), false);
 
-                for (String url : saveImage(dto, multipartFileList)) {
-                    sb = new StringBuilder();
-                    sb.append("images/").append(dto.getEntity()).append("/").append(dto.getId()).append("/");
-
-                    S3Image currentMemberS3Image = currentMember.getS3Image();
-                    currentMemberS3Image.setFilePath(url);
-                    currentMemberS3Image.setFileName(url.split(sb.toString())[1]);
-                    currentMemberS3Image.setFileRepo(sb.toString());
-
+                // Member의 S3Image는 삭제 후 재등록이 아닌 수정하는 방식
+                for (S3Image s3Image : makeNewS3ImageObject(dto, multipartFileList)) {
+                    currentMember.getS3Image().setFilePath(s3Image.getFilePath());
+                    currentMember.getS3Image().setFileName(s3Image.getFileName());
+                    currentMember.getS3Image().setFileRepo(s3Image.getFileRepo());
                     memberService.forceAuthentication(currentMember);
                 }
             }
+            // expenses : 단일 파일 업로드
+            case "expenses" -> {
+                Expenses currentExpenses = expensesRepository.findById(dto.getId())
+                        .orElseThrow(() -> new IdNotFoundException("해당하는 id가 존재하지 않습니다."));
+                S3Image newS3Image = makeNewS3ImageObject(dto, multipartFileList).get(0);
+                newS3Image.setExpenses(currentExpenses);
+                currentExpenses.setS3Image(newS3Image);
+            }
         }
     }
+
+    public List<S3Image> makeNewS3ImageObject(S3ImageDto dto, List<MultipartFile> multipartFileList) {
+        List<S3Image> imageList = new ArrayList<>();
+        StringBuilder sb;
+        for (String url : saveImage(dto, multipartFileList)) {
+            sb = new StringBuilder();
+            sb.append("images/").append(dto.getEntity()).append("/").append(dto.getId()).append("/");
+            S3Image savedImage = s3ImageRepository.save(S3Image.builder()
+                    .filePath(url)
+                    .fileName(url.split(sb.toString())[1])
+                    .fileRepo(sb.toString())
+                    .build());
+            imageList.add(savedImage);
+        }
+        return imageList;
+    }
+
 
     /**
      * 파일 이름을 새로 지정하기 위한 함수
@@ -135,7 +155,7 @@ public class AWSService {
                                 .withCannedAcl(CannedAccessControlList.PublicRead)
                 );
             } catch (IOException e) {
-                log.warn("S3Image Image Upload Fail=()", e);
+                log.warn("S3Image Image Upload Fail={}", e.toString());
             }
 
             String imagePath = amazonS3Client.getUrl(bucketName, sb.toString()).toString(); // 접근가능한 URL 가져오기
@@ -172,10 +192,13 @@ public class AWSService {
 
         log.info("S3 : 삭제할 이미지의 Key = ({}) {}", isExistObject, deleteObjectRequest.getKey());
 
-        if (isExistObject) {
+        if (isExistObject && deleteObjectRequest.getKey().contains("/images/static/profile-image.jpg")) {
+            log.info("S3 : {} 이미지가 삭제되었습니다.", deleteObjectRequest.getKey());
             amazonS3Client.deleteObject(
                     new DeleteObjectRequest(bucketName, deleteObjectRequest.getKey())
             );
+        } else {
+            log.info("S3 : {} 이미지가 삭제되지 않았습니다.", deleteObjectRequest.getKey());
         }
         if (isDelete) {
             s3ImageRepository.delete(s3Image);
